apiRequest: Add proper error handling for requests.
There is a bug in the app. Whenever the API fails to get response, it
does not shows any error message on the screen and the spinner
animations continues infinitely. In this commit, an alert is added which
appears whenever any error occurs so that the user can know about it and
resolve it.