pagination: Added better UI for pagination.
This commit improves the UI of pagination component. Their is a
currently a problem from the Google's Book API. The "totalItems"
property of API seems to be buggy as when we make request to it with
"startIndex" parameter, it gives diffrent values which are not
consistent with the searchIndex number.